+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -6,10 +6,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.View;
 import webmvc.org.springframework.web.servlet.mvc.AnnotationHandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.AnnotationHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.HandlerAdapters;
+import webmvc.org.springframework.web.servlet.mvc.HandlerMappings;
 
 public class DispatcherServlet extends HttpServlet {
