+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/HandlerMappings.java
@@ -1,7 +1,6 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc;
 
 import jakarta.servlet.http.HttpServletRequest;
-import webmvc.org.springframework.web.servlet.mvc.HandlerMapping;
 
 import java.util.ArrayList;
 import java.util.List;
